extract constants to a file
