+++ src/constants/path.ts
@@ -1,0 +1,4 @@
+export const FIREWORKS_XML_PATH = "fireworks.xml";
+export const FOUNTAIN_TEXTURE_PATH = "public/fountain.png";
+export const PARTICLE_TEXTURE_PATH = "public/particle.png";
+export const ROCKET_TEXTURE_PATH = "public/rocket.png";
